Añadido CRUD de instalación y reserva

+++ backend/database/migrations/2025_07_22_174336_create_instalacion_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
             $table->foreignId('municipio_id')->constrained('municipio')->onDelete('cascade');
             $table->string('nombre');
             $table->string('direccion');
-            $table->float('precio');
+            $table->decimal('precio', 5, 2);
             $table->timestamps();
         });
     }
